requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
xgboost>=2.0.0
optuna>=3.3.0
matplotlib>=3.3.0
seaborn>=0.12.0
psutil>=5.9.0
gputil>=1.4.0
joblib>=1.3.0
python-dotenv